Synchronization
This is the first commit which uploads current work to origin.

Currently stable functions:
	auto_print.py
	get_asset_stats.py
	merge_helper.py
	range_list_gen.py

For confidentiality, secret.py and configs.py won't be included.
Since configs.py is complexer, configs_default.py is offered,
while secret.py should be created by the user. Only three variables
should be defined in secret.py:
	username_default, passowrd_default, password_asset

get_VIP.py is in test stage, so it may be updated soon in the
following commits.