get_VIP.py: finding troublesome debtors.
Here a VIP means a person who owes lots of money or cases.
The process flow is following:
	1. Scan all cases between 103y to 107y.
	2. Suppose we've got UIDs of all debtors' of the
		department from 1.
	3. Use those UIDs to get everyone's debt statistics.
	4. Sort them by number of cases or amount of money,
		depanding on user's command.
	5. Print out the cases of those people for merging cases.

On the way to this file, some bugs in func_lib.py were found and
fixed. Besides, get_case_stats is updated due to some changes of
the system. Tested well for department of 孝 on 2018/06/27.