get_clear_cases.py: finding end-able cases.
The criterion is unpaid amount of money. get_topay_summary() is
frequently used in this program, so it is highly suggested that
users shouldn't set a wide range of years to search.

Recently it is found that the system will reject/close connection
under heavy loading, so an exception is set to avoid such
condition, with the current progress reported on the screen.