get_VIP.py: error handling added.
This script usually leads to heavy work for the server, and this
sometimes causes some errors during the process, which makes all
previous done work in vain.

To reduce the loss, if get_case_stats() raises an error, it will
retry the operation for two times (so at most three times totally)
before printing alert messages.