Supporting wholly-owned company functions.
If a company is wholly owned by the owner, the owner's asset should
be listed in the asset page as well. To check whether one company
is wholly-owned, get_case_details() is modified to return more
information including DUTY_NAME, DUTY_IDNO, whether it is wholly-
owned, and the executaion begin date is also returned for further
usage in the future.

Besides, get_asset_stats.py is modified to use rough estimation
for undo amount to boost the performance.