get_asset_stats.py: showing wholly-owned mark.
Though wholly-owned cases can be detected, only UID list will be
different in this script previously. If a case is wholly-owned or
partnership, it should be marked in front of the DUTY_NAME as it
is in auto_read.py.

Since all information is got from get_case_details(),
get_duty_name_and_uid() is abandoned for better readability.