merge_helper.py: adding pooling case list.
It is a good idea to first take out all the cases (by a list of
all cases to take), and then merge the cases by the original list.

The first step is much faster since no need to walk around anymore.
The second step is not slow as well because it is much easier to
find a case in a little scale sorted pile of cases.

We need three arguments now, including one input file and two
output files for the two lists. For compatibility, the third
argument is optional, so if the third argument is not given, it
will behave as before.