auto_read.py: supporting UID input files.
Since situation list is based on a case rather than a debtor, there
must be a strategy to determine which cases should be read for the
debtor. A naive method is to read all cases of the debtors, which
will lead to performance tradedy. Here the following rule is used:
	1. Find the oldest case of 01, 03 and 04 (w/o LI cases) for
	    the debtor.
	2. Look up those cases and pool the results.
	3. If no commands are done, raise a warning message which
	    tells the user that the main execution case is
	    undefined and not needed.
	4. Else pick the latest command from the pool, and pick the
	    case of the command as the main execution case.
	5. Look up the main execution case.

Overall at most four cases will be looked up, so the performance
should be acceptable.