get_asset_stats.py: handling health insurance.
Health insurance is a special source of property because it is
single source property. That is, no matter what the source unit is,
all health insurance data should be treated as if they are from an
identical source. So it should be separated from other data, which
use (income_description, unit_name, idno_of_bank) as the feature.