share_lib.py: a library for all shared functions.
To make func_lib.py even simpler, it should contain only functions
that communicate with the server, and sub-functions that they need.
Other functions that more than one scripts call should be collected
to some library but not func_lib. So a new file, share_lib.py, is
created to be the one.

At the same time, previously a script may call a function defined
in another script, which potentially leads to "incomplete update"
problem because nobody remebers to fix those files which reference
current modified files. So if a function is shared by more than
one script, such function should be put into share_lib.py for
long-term maintenance.