merge_helper.py: allowing evidence even if cleared.
In some cases, the numbers given from the system may be wrong. That
is, those numbers may tell that the cases have been cleared, but
in fact the cases still need evidence. Previously such cases will
fail to end even if the user assign the ending situation manually.

Now has_cleared=False is not an essential condition for evidence
anymore, and so that the cases above will be detected due to
possible ambiguity of the ending situation, and the cases can be
ended by manually assignment.