func_lib.py: get_ended_case_stats implemented.
In the previous version, this function is named get_old_cases_stats
because it was only used to get very old cases (before 98y). But
in fact the page supports all kinds of cases, and now it is completed.

Notice that if the user passes nothing to this function, the results
will be an alert page, which will lead to an error. However, since
this is a meaningless operation, and it is a trouble to handle such
case, no error handling is prepared for the case.