Encoding for Windows: use UTF8 with BOM.
Most devices in Enforcement Agency are using Windows system, which
takes cp950 as the default encoding method. That's to say, any output
files that doesn't include BOM will be opened with cp950 by Excel,
while the real encoding method is UTF8. To avoid this problem, BOM
should be used at the head of the files.