max_case_seqno: a new variable in configs.
As coming cases are getting more and more, it is hard to determine
a constant number which will keep larger than maximal sequence
number, but not too large to do too many useless operations.

So here we set it as a variable in configs.py and is shared by all
applications that need such number to keep it easy to change and
consistent everywhere.