auto_read.py: main-case query for both input.
In the previous version, input consisting of case seq-nums will raise
queries for the numbers directly. However, it is possible that someone
use a case to represent a debtor, which is similar to UID. In such case,
the case number should be used to find the main case number and then
start the query.

Besides, some enhancement for user experience is added.
    1. The error message becomes clearer.
    2. Progress will be showed on the screen similar to other scripts.