get_todestroy_case_stats.py: abandoned.
This script is fine logically, but its background page (TPKH0401F)
is full of bugs and so that it can't give us right output until the
system is fixed. On the other hand, we found that there are other
ways to attain the same results without handling those bugs, so we
will use one of them to replace this script.