func_lib.py: login_asset() refactored.
It is strange to allow a function print out the error message by
itself because the caller will lose its control to output content
and the chance to handle the error.

When there are no errors, no message is returned. To make it specific,
if there are no errors, True will be returned; otherwise, error
message will be returned.