get_todestroy_stats.py: as file name.
This script is developed to replace get_todestroy_case_stats.py,
which has been abandoned in previous commit. The script must be based
on the output file of get_ended_case_list.py because it doesn't access
the server by itself, but just reads the output file and generates
output according to the content.

Two kinds of files will be generated:
    1. The to-destroy case list of each department.
    2. A .csv file which is the overall statistics.