ranged_case_list(): checking whether it is clear.
In the previous version, this function will only check whether it is
health/labor insurance. Now it will check whether the case is clear.
Those cases which are clear will be marked by '#'.

For a range of cases, it will be marked as clear cases only if all
cases in the range are clear.

Besides, there may be a rare condition: the debtor has a non-LI case
whose number is n, and a range of cases whose number are n+1~m. In
such cases, it was possible that the script will treat them all as
a range of cases n~m. In this commit, such conditions are handled
as well.